Allow opt-in parsing of Cython (.pyx) files
Adds CLI argument `--force-pyx` to force the parsing of Cython
files (ending in .pyx). By default only .py files are accepted.

Although the Cython specific syntax is not handled, some .pyx files
only contain regular python code. Whether this is known or not will be
at the users discretion, which is why this feature is explicitly
opt-in. By opting in the user takes responsibility for any parsing issues.